api/requirements.txt
fastapi
uvicorn[standard]
pandas
requests
numpy
pyarrow
google-cloud-storage
google-cloud-bigquery
gcsfs
db-dtypes
